Split parser, improved lexer and fixed compiler.
Parser was split to lexer and compiler.
Lexer now allows stay HTML without breaking.
Compiler fixed to emit proper literals (stay injects solved).
